Refactor fixtures location in storage nfs tests
