+++ openvair/modules/storage/tests/conftest.py
@@ -4,30 +4,17 @@ from typing import Any, Dict, Generator
 from pathlib import Path
 
 import pytest
-from fastapi import status
 from fastapi.testclient import TestClient
 
 from openvair.libs.log import get_logger
 from openvair.libs.testing.utils import (
     create_resource,
-    delete_resource,
     cleanup_partitions,
-    cleanup_all_volumes,
     get_disk_partitions,
     cleanup_all_storages,
-    wait_for_field_value,
-    cleanup_all_templates,
-    generate_test_entity_name,
 )
 from openvair.libs.testing.config import storage_settings
-from openvair.modules.template.shared.enums import TemplateStatus
-from openvair.modules.volume.entrypoints.schemas import CreateVolume
 from openvair.modules.storage.entrypoints.schemas import CreateLocalPartition
-from openvair.modules.volume.service_layer.services import VolumeStatus
-from openvair.modules.storage.service_layer.services import StorageStatus
-from openvair.modules.template.entrypoints.schemas.requests import (
-    RequestCreateTemplate,
-)
 
 LOG = get_logger(__name__)
 
@@ -83,109 +70,3 @@ def local_partition(
     )
 
     yield partition
-
-
-@pytest.fixture(scope='function')
-def nfs_storage(client: TestClient) -> Generator[Dict, None, None]:
-    """Creates a test NFS storage and deletes it after test."""
-    cleanup_all_templates()
-    cleanup_all_volumes()
-    cleanup_all_storages()
-
-    nfs_storage_data = {
-        'name': generate_test_entity_name('nfs_storage'),
-        'description': 'Test NFS storage for integration tests',
-        'storage_type': 'nfs',
-        'specs': {
-            'ip': str(storage_settings.storage_nfs_ip),
-            'path': str(storage_settings.storage_nfs_path),
-            'mount_version': '4',
-        },
-    }
-
-    storage_nfs = create_resource(
-        client,
-        '/storages/create/',
-        nfs_storage_data,
-        'nfs_storage',
-    )
-
-    wait_for_field_value(
-        client=client,
-        path=f"/storages/{storage_nfs['id']}/",
-        field='status',
-        expected=StorageStatus.available.name,
-    )
-
-    yield storage_nfs
-
-    cleanup_all_volumes()
-    cleanup_all_templates()
-
-    delete_response = client.delete(f"/storages/{storage_nfs['id']}/delete")
-    if delete_response.status_code != status.HTTP_202_ACCEPTED:
-        LOG.warning(
-            (
-                f'Failed to delete test storage: {delete_response.status_code},'
-                f' {delete_response.text}'
-            )
-        )
-
-
-@pytest.fixture(scope='function')
-def nfs_volume(
-    client: TestClient, nfs_storage: Dict
-) -> Generator[Dict, None, None]:
-    """Creates a test volume on NFS storage and deletes it after each test."""
-    volume_data = CreateVolume(
-        name=generate_test_entity_name('volume'),
-        description='Volume for integration tests on NFS storage',
-        storage_id=nfs_storage['id'],
-        format='qcow2',
-        size=1024,  # 1GB
-        read_only=False,
-    ).model_dump(mode='json')
-
-    volume_nfs = create_resource(
-        client, '/volumes/create/', volume_data, 'nfs_volume'
-    )
-
-    wait_for_field_value(
-        client=client,
-        path=f'/volumes/{volume_nfs["id"]}/',
-        field='status',
-        expected=VolumeStatus.available.name,
-    )
-
-    yield volume_nfs
-
-    delete_resource(client, '/volumes', volume_nfs['id'], 'nfs_volume')
-
-
-@pytest.fixture(scope='function')
-def nfs_template(
-    client: TestClient, nfs_storage: Dict, nfs_volume: Dict
-) -> Generator[Dict, None, None]:
-    """Creates a test template on NFS storage and deletes it after each test."""
-    template_data = RequestCreateTemplate(
-        base_volume_id=nfs_volume['id'],
-        name=generate_test_entity_name('template'),
-        description='Template for integration tests on NFS storage',
-        storage_id=nfs_storage['id'],
-        is_backing=False,
-    ).model_dump(mode='json')
-
-    template_nfs = create_resource(
-        client, '/templates/', template_data, 'nfs_template'
-    )['data']
-
-    wait_for_field_value(
-        client=client,
-        path=f"/templates/{template_nfs['id']}/",
-        field='status',
-        expected=TemplateStatus.AVAILABLE,
-    )
-
-    yield template_nfs
-
-    delete_resource(client, '/templates', template_nfs['id'], 'nfs_template')
